requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
google-generativeai==0.7.1
assemblyai==0.24.0
murf==0.4.0
pydantic==2.7.4